Proof of concept with vertex array objects
